+++ src/main/kotlin/com/zorbeytorunoglu/thebestranks/configuration/ranks/Requirement.kt
@@ -1,6 +1,6 @@
 package com.zorbeytorunoglu.thebestranks.configuration.ranks
 
-class Requirement(private val rank: Rank,private val placeholder: String, private val required: Int, private val denyMessage: String, private val guiMessage: String) {
+class Requirement(private val rank: Rank,private val placeholder: String, private val required: Double, private val denyMessage: String, private val guiMessage: String) {
 
     fun getRank(): Rank {
         return rank
@@ -10,7 +10,7 @@ class Requirement(private val rank: Rank,private val placeholder: String, privat
         return placeholder
     }
 
-    fun getRequired(): Int {
+    fun getRequired(): Double {
         return required
     }
